Move Pozition functionality to dedicated folder
In preparation for adding more functionality. I don't want the index
file to be cluttered.
